Refactor parsing methods out of DataSource
