+++ tests/Searchlight.Tests/ExceptionTests.cs
@@ -29,7 +29,7 @@ namespace Searchlight.Tests
             var originalFilter = "((()))";
             var ex = Assert.ThrowsException<EmptyClause>((Action)(() =>
             {
-                var query = src.Parse(originalFilter);
+                var query = src.ParseFilter(originalFilter);
                 var sql = query.ToSqlServerCommand();
             }));
             Assert.AreEqual(originalFilter, ex.OriginalFilter);
