+++ tests/Searchlight.Tests/SqlExecutorTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+// ReSharper disable StringLiteralTypo
 
 namespace Searchlight.Tests
 {
@@ -35,14 +36,14 @@ namespace Searchlight.Tests
                 "b in (1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140, 141, 142, 143, 144, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 158, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 171, 172, 173, 174, 175, 176, 177, 178, 179, 180, 181, 182, 183, 184, 185, 186, 187, 188, 189, 190, 191, 192, 193, 194, 195, 196, 197, 198, 199, 200, 201, 202, 203, 204, 205, 206, 207, 208, 209, 210, 211, 212, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222, 223, 224, 225, 226, 227, 228, 229, 230, 231, 232, 233, 234, 235, 236, 237, 238, 239, 240, 241, 242, 243, 244, 245, 246, 247, 248, 249, 250, 251, 252, 253, 254, 255, 256)";
             var ex = Assert.ThrowsException<TooManyParameters>(() =>
             {
-                var query = _source.Parse(originalFilter);
+                var query = _source.ParseFilter(originalFilter);
                 query.ToSqlServerCommand();
             });
             Assert.AreEqual(originalFilter, ex.OriginalFilter);
 
             // Verify that 0 is treated as unlimited
             _source.MaximumParameters = 0;
-            var syntax = _source.Parse(originalFilter);
+            var syntax = _source.ParseFilter(originalFilter);
             Assert.IsNotNull(syntax);
             _source.MaximumParameters = null;
 
@@ -53,7 +54,7 @@ namespace Searchlight.Tests
             };
             ex = Assert.ThrowsException<TooManyParameters>(() =>
             {
-                var query = _source.Parse(originalFilter);
+                var query = _source.ParseFilter(originalFilter);
                 query.ToSqlServerCommand();
             });
             Assert.AreEqual(originalFilter, ex.OriginalFilter);
@@ -63,25 +64,25 @@ namespace Searchlight.Tests
         public void CaseSensitivity()
         {
             // First test, lowercase
-            var query = _source.Parse("b between 1 and 5");
+            var query = _source.ParseFilter("b between 1 and 5");
             var sql = query.ToSqlServerCommand();
             Assert.AreEqual("b BETWEEN @p1 AND @p2", sql.WhereClause.ToString());
             Assert.AreEqual(2, sql.Parameters.Count);
 
             // Second test, proper case
-            query = _source.Parse("b Between 1 And 5");
+            query = _source.ParseFilter("b Between 1 And 5");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("b BETWEEN @p1 AND @p2", sql.WhereClause.ToString());
             Assert.AreEqual(2, sql.Parameters.Count);
 
             // Third test, uppercase
-            query = _source.Parse("b BETWEEN 1 AND 5");
+            query = _source.ParseFilter("b BETWEEN 1 AND 5");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("b BETWEEN @p1 AND @p2", sql.WhereClause.ToString());
             Assert.AreEqual(2, sql.Parameters.Count);
             
             // Fourth test, inverse
-            query = _source.Parse("b not BETWEEN 1 AND 5");
+            query = _source.ParseFilter("b not BETWEEN 1 AND 5");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("b NOT BETWEEN @p1 AND @p2", sql.WhereClause.ToString());
             Assert.AreEqual(2, sql.Parameters.Count);
@@ -141,7 +142,7 @@ namespace Searchlight.Tests
         [TestMethod]
         public void FilterParseTest()
         {
-            var query = _source.Parse("a = 'booya' AND b != 1");
+            var query = _source.ParseFilter("a = 'booya' AND b != 1");
             var sql = query.ToSqlServerCommand();
             Assert.AreEqual("a = @p1 AND b <> @p2", sql.WhereClause.ToString());
             Assert.AreEqual(2, sql.Parameters.Count);
@@ -152,22 +153,22 @@ namespace Searchlight.Tests
         [TestMethod]
         public void NullInWhereClause()
         {
-            var query = _source.Parse("a is null");
+            var query = _source.ParseFilter("a is null");
             var sql = query.ToSqlServerCommand();
             Assert.AreEqual("a IS NULL", sql.WhereClause.ToString());
             Assert.AreEqual(0, sql.Parameters.Count);
 
-            query = _source.Parse("a is not null");
+            query = _source.ParseFilter("a is not null");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("a IS NOT NULL", sql.WhereClause.ToString());
             Assert.AreEqual(0, sql.Parameters.Count);
 
-            query = _source.Parse("(  a  is  not  null )  or   ( a  is  null  )  ");
+            query = _source.ParseFilter("(  a  is  not  null )  or   ( a  is  null  )  ");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("(a IS NOT NULL) OR (a IS NULL)", sql.WhereClause.ToString());
             Assert.AreEqual(0, sql.Parameters.Count);
 
-            query = _source.Parse("(((  a  is  not  null ))  or   ( a  is  null  ))  ");
+            query = _source.ParseFilter("(((  a  is  not  null ))  or   ( a  is  null  ))  ");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("(((a IS NOT NULL)) OR (a IS NULL))", sql.WhereClause.ToString());
             Assert.AreEqual(0, sql.Parameters.Count);
@@ -175,7 +176,7 @@ namespace Searchlight.Tests
 
         public string ParseWhereClause(string filter)
         {
-            var query = _source.Parse(filter);
+            var query = _source.ParseFilter(filter);
             var sql = query.ToSqlServerCommand();
             return sql.WhereClause.ToString();
         }
@@ -220,40 +221,40 @@ namespace Searchlight.Tests
         public void TypeComparisons()
         {
             // Test the Int64
-            var query = _source.Parse("collong eq 123456789123456");
+            var query = _source.ParseFilter("collong eq 123456789123456");
             var sql = query.ToSqlServerCommand();
             Assert.AreEqual("colLong = @p1", sql.WhereClause.ToString());
             Assert.AreEqual(123456789123456, sql.Parameters["@p1"]);
 
             // Test the guid
-            query = _source.Parse(String.Format("colguid eq '{0}'", Guid.Empty.ToString()));
+            query = _source.ParseFilter(String.Format("colguid eq '{0}'", Guid.Empty.ToString()));
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("colGuid = @p1", sql.WhereClause.ToString());
             Assert.AreEqual(Guid.Empty, sql.Parameters["@p1"]);
 
             // Test the nullable guid
-            query = _source.Parse(String.Format("colNullableGuid is null or colNullableGuid = '{0}'",
+            query = _source.ParseFilter(String.Format("colNullableGuid is null or colNullableGuid = '{0}'",
                 Guid.Empty.ToString()));
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("colNullableGuid IS NULL OR colNullableGuid = @p1", sql.WhereClause.ToString());
             Assert.AreEqual(Guid.Empty, sql.Parameters["@p1"]);
 
             // Test the ULONG and nullable ULONG
-            query = _source.Parse("colULong > 12345 or colNullableULong = 6789456");
+            query = _source.ParseFilter("colULong > 12345 or colNullableULong = 6789456");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("colULong > @p1 OR colNullableULong = @p2", sql.WhereClause.ToString());
             Assert.AreEqual(12345UL, sql.Parameters["@p1"]);
             Assert.AreEqual(6789456UL, sql.Parameters["@p2"]);
 
             // Test the ULONG and nullable ULONG when compared to a boolean - necessary for redshift
-            query = _source.Parse("colULong = true OR colULong = false");
+            query = _source.ParseFilter("colULong = true OR colULong = false");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("colULong = @p1 OR colULong = @p2", sql.WhereClause.ToString());
             Assert.AreEqual(1UL, sql.Parameters["@p1"]);
             Assert.AreEqual(0UL, sql.Parameters["@p2"]);
 
             // Nullable variant
-            query = _source.Parse("colNullableULong = true OR colNullableULong = false");
+            query = _source.ParseFilter("colNullableULong = true OR colNullableULong = false");
             sql = query.ToSqlServerCommand();
             Assert.AreEqual("colNullableULong = @p1 OR colNullableULong = @p2", sql.WhereClause.ToString());
             Assert.AreEqual(1UL, sql.Parameters["@p1"]);
@@ -264,7 +265,7 @@ namespace Searchlight.Tests
         public void SqlServerBasicQuery()
         {
             // Basic query including where and order
-            var query = _source.Parse("collong eq 123456789123456", null, "b ascending");
+            var query = _source.Parse(new FetchRequest() { filter = "collong eq 123456789123456", order = "b ascending" });
             _engine.useResultSet = false;
             var sql = query.ToSqlServerCommand();
             Assert.AreEqual("SET NOCOUNT ON;\nSET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;\n" +
